moved AbstractJobConfigurationBuilder to de.uka.ipd.sdq.workflow.launchconfig.core
