+++ bundles/de.uka.ipd.sdq.workflow.launchconfig.core/META-INF/MANIFEST.MF
@@ -3,7 +3,8 @@ Bundle-ManifestVersion: 2
 Bundle-Name: Run Config Core
 Bundle-SymbolicName: de.uka.ipd.sdq.workflow.launchconfig.core
 Bundle-Version: 5.2.1.qualifier
-Export-Package: de.uka.ipd.sdq.workflow.launchconfig.core
+Export-Package: de.uka.ipd.sdq.workflow.launchconfig.core,
+ de.uka.ipd.sdq.workflow.launchconfig.core.configbuilder
 Require-Bundle: org.eclipse.debug.core;bundle-version="3.20.0",
  de.uka.ipd.sdq.workflow;bundle-version="5.2.1"
 Bundle-ActivationPolicy: lazy
