+++ bundles/de.uka.ipd.sdq.workflow.launchconfig.core/src/de/uka/ipd/sdq/workflow/launchconfig/core/configbuilder/AbstractJobConfigurationBuilder.java
@@ -1,4 +1,5 @@
-package de.uka.ipd.sdq.workflow.launchconfig.configbuilder;
+/** */
+package de.uka.ipd.sdq.workflow.launchconfig.core.configbuilder;
 
 import java.util.Map;
