+++ bundles/de.uka.ipd.sdq.workflow.launchconfig/META-INF/MANIFEST.MF
@@ -21,7 +21,6 @@ Require-Bundle: org.eclipse.ui,
 Bundle-ActivationPolicy: lazy
 Bundle-Vendor: sdq.ipd.uka.de
 Export-Package: de.uka.ipd.sdq.workflow.launchconfig,
- de.uka.ipd.sdq.workflow.launchconfig.configbuilder,
  de.uka.ipd.sdq.workflow.launchconfig.extension,
  de.uka.ipd.sdq.workflow.launchconfig.tabs
 Bundle-RequiredExecutionEnvironment: JavaSE-17
