--- bundles/de.uka.ipd.sdq.workflow.launchconfig/src/de/uka/ipd/sdq/workflow/launchconfig/configbuilder/package-info.java
@@ -1,11 +1,0 @@
-/**
- * The configuration builder package of the workflow engine.
- * 
- * This package provides the generic infrastructure to define builder support
- * the creation of configuration objects.
- *
- * @author Palladio Project
- * @version 1.0
- *
-*/
-package de.uka.ipd.sdq.workflow.launchconfig.configbuilder;
