+++ bundles/de.uka.ipd.sdq.workflow.launchconfig/src/de/uka/ipd/sdq/workflow/launchconfig/tabs/AbstractConfigBuilderTab.java
@@ -5,7 +5,7 @@ import org.eclipse.debug.core.ILaunchConfiguration;
 import org.eclipse.debug.ui.AbstractLaunchConfigurationTab;
 
 import de.uka.ipd.sdq.workflow.configuration.AbstractJobConfiguration;
-import de.uka.ipd.sdq.workflow.launchconfig.configbuilder.AbstractJobConfigurationBuilder;
+import de.uka.ipd.sdq.workflow.launchconfig.core.configbuilder.AbstractJobConfigurationBuilder;
 
 /**
  * The Class AbstractConfigBuilderTab.
